PR1: Add admin layers to params (#11)
* Add admin layers to params

* PR2: Add layer params to the deeplink (#12)

* renames and move dirs

* Move helpers and use NRW name

* clean up and make params hook

* fix name and comments

* Potential fix for pull request finding


* Potential fix for pull request finding


* pr feedback

* pr4: Rename to NRW from IBF (#14)

* renames and move dirs

* Move helpers and use NRW name

* Add layer parms

* apply pr feedback

* P3: Clean up helpers and param handlers (#13)

* renames and move dirs

* Move helpers and use NRW name

* clean up and make params hook

* fix name and comments

* Potential fix for pull request finding


* Potential fix for pull request finding


* pr feedback

* pr4: Rename to NRW from IBF (#14)

* renames and move dirs

* Move helpers and use NRW name

---------


---------


---------
