+++ README_IBF_addendum.md
@@ -1,4 +1,4 @@
-# Readme addendum for IBF
+# Readme addendum for NRW
 
 Some of the [readme](./README.md) instructions don't work or are incomplete. Rather than writing over the original readme and risk PRs getting held up over comments on strings, use this as the troubleshooting addendum.
