+++ app/src/utils/nrw/nrwConstants.ts
@@ -1,0 +1,13 @@
+// Map property strings
+export const noCountrySelectedValue = 'None';
+
+// Default map values
+export const defaultMapZoom = 3;
+
+// Data field keys, for instance keys in the GeoJSON data.
+export const COUNTRY_FIELD_KEY = 'country';
+export const PLACE_CODE_FIELD_KEY = 'code';
+export const ADMIN_LEVEL_FIELD_KEY = 'admin_level';
+export const ADMIN1_PCODE_FIELD_KEY = 'admin1_pcode';
+export const ADMIN2_PCODE_FIELD_KEY = 'admin2_pcode';
+export const ADMIN3_PCODE_FIELD_KEY = 'admin3_pcode';
